ProgramacionMultimediaYDispositivosMoviles/examenes/examen02/unbroken/src/dependencies.py
import jwt
from fastapi import Security, HTTPException
from fastapi.security import OAuth2PasswordBearer

SECRET = "secret"
ALGO = "HS256"
oauth2 = OAuth2PasswordBearer(tokenUrl="/auth/login")


def get_current_user(token: str = Security(oauth2)):
    try:
        payload = jwt.decode(token, SECRET, algorithms=[ALGO])
        return payload.get("sub")
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")
