ProgramacionMultimediaYDispositivosMoviles/examenes/examen02/unbroken/pyproject.toml
[project]
name = "unbroken"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "freezegun>=1.5.5",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
